I keep forgetting those values are stored in Longs.  Is this kosher?  Just casting?  I'm really just trying to avoid handling the null case (why can't we handle unboxing Longs into longs when we transmit them to the AS side?  This really is a nuisance)

+++ src/java/com/threerings/msoy/badge/data/all/EarnedBadge.java
@@ -47,7 +47,7 @@ public class EarnedBadge extends Badge
         if (o instanceof EarnedBadge) {
             EarnedBadge other = (EarnedBadge)o;
             return super.equals(other) && other.level == this.level &&
-                other.whenEarned == this.whenEarned;
+                ((long)other.whenEarned) == ((long)this.whenEarned);
         }
         return false;
     }
